requirements.txt
openai
serpapi
python-dotenv
tqdm
pinecone-client
langchain
bigtree